requirements.txt
allure-pytest==2.8.17
allure-python-commons==2.8.17
atomicwrites==1.4.0
attrs==19.3.0
chromedriver==2.24.1
colorama==0.4.3
enum34==1.1.10
importlib-metadata==1.7.0
iniconfig==1.0.1
jsonpickle==1.4.1
lxml==4.5.2
more-itertools==8.4.0
namedlist==1.7
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
selenium==3.141.0
six==1.15.0
toml==0.10.1
urllib3==1.25.10
zipp==3.1.0
